Integrate qt/3.0 -> qt/main
[git-p4: depot-paths = "//depot/qt/main/": change = 50383]

+++ src/tools/quuid.cpp
@@ -60,7 +60,7 @@
 */
 
 /*!
-  \fn QUuid::QUuid( uint l, ushort w1, ushort w2, uchar b1, uchar b2, uchar b3, uchar b4, uchar b5, uchar b6, uchar b7, uchar b8 )
+  \fn QUuid::QUuid( ulong l, ushort w1, ushort w2, uchar b1, uchar b2, uchar b3, uchar b4, uchar b5, uchar b6, uchar b7, uchar b8 )
 
   Creates an UUID with the value specified by the parameters, \a l, \a
   w1, \a w2, \a b1, \a b2, \a b3, \a b4, \a b5, \a b6, \a b7, \a b8.
